implement authentication.

+++ src/main/java/com/abivan/graphics/repository/GraphicRepository.java
@@ -3,5 +3,5 @@ package com.abivan.graphics.repository;
 import com.abivan.graphics.domain.Graphic;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GraphicRepository extends JpaRepository<Graphic, Integer> {
+public interface GraphicRepository extends JpaRepository<Graphic, Long> {
 }
